+++ Console/Console.csproj
@@ -7,6 +7,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="CliFx" Version="2.3.5" />
     <PackageReference Include="Concentus" Version="2.2.2" />
     <PackageReference Include="Concentus.Native" Version="1.5.2" />
     <PackageReference Include="Nito.AsyncEx" Version="5.1.2" />
