+++ Readme.md
@@ -4,7 +4,7 @@ Listen to youtube music from the terminal
 ![imagen](https://github.com/user-attachments/assets/2be7eb04-6baa-43b0-938d-8a6354a2f387)
 
 ## TODO
-- [ ] Implement login with cookies (either via browser or by manually grabbing them)
+- [X] Implement login with cookies (either via browser or by manually grabbing them)
 - [X] Implement Youtube music suggestions
 - [X] Repeat song or list of songs
 - [X] Seek time with only keyboard
@@ -19,6 +19,10 @@ Listen to youtube music from the terminal
 - Almost no CPU usage 
 - Small memory usage
 - No external dependencies (No MPV, no yt-dlp or youtube-dl)
+
+## Login
+To login you can extract the cookies from music.youtube.com following this [instructions](https://www.reddit.com/r/youtubedl/wiki/cookies/)
+And then run
 
 ## Download
 > [!WARNING]\
